+++ libraries/Wire/src/Wire.h
@@ -43,12 +43,10 @@ class TwoWire : public Stream, public IOMaster {
     void end();
     void setClock(uint32_t);
 
-    void beginTransmission(uint8_t);
-    uint8_t endTransmission(bool stopBit);
-    uint8_t endTransmission(void);
+    void beginTransmission(uint8_t address);
+    uint8_t endTransmission(bool stopBit = true);
 
-    uint8_t requestFrom(uint8_t address, size_t quantity, bool stopBit);
-    uint8_t requestFrom(uint8_t address, size_t quantity);
+    uint8_t requestFrom(uint8_t address, size_t quantity, bool stopBit = true);
 
     size_t write(uint8_t data);
     size_t write(const uint8_t * data, size_t quantity);
@@ -73,11 +71,11 @@ class TwoWire : public Stream, public IOMaster {
     ap3_gpio_pin_t  _padSCL;
 
     bool _transmissionBegun;
-    uint8_t _txAddress;
+    uint8_t _transmissionAddress;
 
     
     RingBufferN<AP3_WIRE_RX_BUFFER_LEN> _rxBuffer;// RX Buffer    
-    RingBufferN<AP3_WIRE_TX_BUFFER_LEN> _txBuffer;//TX buffer
+    RingBufferN<AP3_WIRE_TX_BUFFER_LEN> _txBuffer;// TX buffer
     uint8_t _linearBugger[AP3_WIRE_LINEAR_BUFFER_LEN]; // ToDo: choose a more efficient way to handle this
     uint8_t txAddress;